+++ R/options.R
@@ -40,6 +40,26 @@ rcmd_process_options <- function(...) {
   update_options(rcmd_process_options_default(), ...)
 }
 
+#' Create options for an [rscript_process] object
+#'
+#' @param ... Options to override, named arguments.
+#' @return A list of options.
+#'
+#' `rscript_process_options()` creates a set of options to initialize a new
+#' object from the `rscript_process` class. Its arguments must be named,
+#' the names are used as option names. The options correspond to (some of)
+#' the arguments of the [rscript()] function. At least the `script` option
+#' must be specified, the script file to run.
+#'
+#' @export
+#' @examples
+#' ## List all options and their default values:
+#' rscript_process_options()
+
+rscript_process_options <- function(...) {
+  update_options(rscript_process_options_default(), ...)
+}
+
 r_process_options_default <- function() {
   list(
     func = NULL,
@@ -48,6 +68,7 @@ r_process_options_default <- function() {
     repos = c(getOption("repos"), CRAN = "https://cloud.r-project.org"),
     stdout = "|",
     stderr = "|",
+    poll_connection = TRUE,
     error = getOption("callr.error", "error"),
     cmdargs = c("--slave", "--no-save", "--no-restore"),
     system_profile = FALSE,
@@ -65,6 +86,7 @@ rcmd_process_options_default <- function() {
     libpath = .libPaths(),
     stdout = "|",
     stderr = "|",
+    poll_connection = TRUE,
     repos = c(getOption("repos"), CRAN = "https://cloud.r-project.org"),
     system_profile = FALSE,
     user_profile = FALSE,
@@ -74,13 +96,28 @@ rcmd_process_options_default <- function() {
   )
 }
 
-#' @importFrom utils modifyList
+rscript_process_options_default <- function() {
+  list(
+    script = NULL,
+    cmdargs = character(),
+    libpath = .libPaths(),
+    stdout = "|",
+    stderr = "|",
+    poll_connection = TRUE,
+    repos = c(getOption("repos"), CRAN = "https://cloud.r-project.org"),
+    system_profile = FALSE,
+    user_profile = FALSE,
+    env = rcmd_safe_env(),
+    wd = ".",
+    color = FALSE
+  )
+}
 
 update_options <- function(old_opts, ...) {
   new_opts <- list(...)
   stopifnot(is.named(new_opts))
   check_for_option_names(old_opts, new_opts)
-  modifyList(old_opts, new_opts)
+  utils::modifyList(old_opts, new_opts)
 }
 
 check_for_option_names <- function(old, new) {