+++ R/script.R
@@ -30,13 +30,17 @@ make_vanilla_script_expr <- function(expr_file, res, error,
 
   message <- function() {
     substitute({
-      msg <- paste0("base64::", base64enc::base64encode(serialize(e, NULL)))
+      if (is.null(e$code)) e$code <- "301"
+      msg <- paste0("base64::", processx::base64_encode(serialize(e, NULL)))
       data <- paste(e$code, msg, "\n")
       con <- processx::conn_create_fd(3, close = FALSE)
       while (1) {
         data <- processx::conn_write(con, data)
         if (!length(data)) break;
         Sys.sleep(.1)
+      }
+      if (!is.null(findRestart("muffleMessage"))) {
+        invokeRestart("muffleMessage")
       }
     })
   }
@@ -74,7 +78,14 @@ make_vanilla_script_expr <- function(expr_file, res, error,
           interrupt = function(e) { `__error__` },
           callr_message = function(e) { `__message__` }
         ),
-        error = function(e) { `__post_hook__`; e },
+
+        ## We need to `stop()` here again, otherwise the error message
+        ## is not printed to stderr. See
+        ## https://github.com/r-lib/callr/issues/80
+        ## However, on R 3.1 and R 3.2 throwing an error here
+        ## will crash the R process. With `try()` the error is still
+        ## printed to stderr, but no real error is thrown.
+        error = function(e) { `__post_hook__`; try(stop(e)) },
         interrupt = function(e) {  `__post_hook__`; e }
       )                                 # nocov end
     },