+++ R/setup.R
@@ -21,6 +21,8 @@ setup_context <- function(options) {
   ## Avoid R CMD check warning...
   repos <- libpath <- system_profile <- user_profile <- load_hook <- NULL
 
+  make_path <- function(x) paste(x, collapse = .Platform$path.sep)
+
   within(options, {
     ## profiles
     profiles <- make_profiles(system_profile, user_profile, repos, libpath,
@@ -28,13 +30,18 @@ setup_context <- function(options) {
     tmp_files <- c(tmp_files, profiles)
 
     ## environment files
-    envs <- make_environ(profiles)
+    envs <- make_environ(profiles, libpath)
     tmp_files <- c(tmp_files, envs)
-    
+
     if (is.na(env["R_ENVIRON"])) env["R_ENVIRON"] <- envs[[1]]
     if (is.na(env["R_ENVIRON_USER"])) env["R_ENVIRON_USER"] <- envs[[2]]
     if (is.na(env["R_PROFILE"])) env["R_PROFILE"] <- profiles[[1]]
     if (is.na(env["R_PROFILE_USER"])) env["R_PROFILE_USER"] <- profiles[[2]]
+
+    ## environment variables
+    if (is.na(env["R_LIBS"])) env["R_LIBS"] <- make_path(libpath)
+    if (is.na(env["R_LIBS_USER"])) env["R_LIBS_USER"] <- make_path(libpath)
+    if (is.na(env["R_LIBS_SITE"])) env["R_LIBS_SITE"] <- make_path(.Library.site)
   })
 }
 
@@ -87,10 +94,17 @@ make_profiles <- function(system, user, repos, libpath, load_hook) {
   c(profile_system, profile_user)
 }
 
-make_environ <- function(profiles) {
+make_environ <- function(profiles, libpath) {
 
   env_sys <- tempfile()
   env_user <- tempfile()
+
+  for (ef in c(env_sys, env_user)) {
+    cat("CALLR_CHILD_R_LIBS=\"${R_LIBS}\"\n",
+        "CALLR_CHILD_R_LIBS_USER=\"${R_LIBS_USER}\"\n",
+        "CALLR_CHILD_R_LIBS_SITE=\"${R_LIBS_SITE}\"\n",
+        file = ef, append = TRUE)
+  }
 
   sys <- Sys.getenv("R_ENVIRON", NA_character_)
   if (is.na(sys)) sys <- file.path(R.home("etc"), "Renviron.site")
@@ -101,13 +115,22 @@ make_environ <- function(profiles) {
   home <- "~/.Renviron"
   if (is.na(user) && file.exists(local)) user <- local
   if (is.na(user) && file.exists(home)) user <- home
-  if (!is.na(user) && file.exists(user))  file.append(env_user, user)
+  if (!is.na(user) && file.exists(user)) file.append(env_user, user)
 
   for (ef in c(env_sys, env_user)) {
     cat("R_PROFILE=\"", profiles[[1]], "\"\n", file = ef,
         append = TRUE, sep = "")
     cat("R_PROFILE_USER=\"", profiles[[2]], "\"\n", file = ef,
         append = TRUE, sep = "")
+    cat("R_LIBS_SITE=\"${CALLR_CHILD_R_LIBS_SITE:-",
+        paste(.Library.site, collapse = .Platform$path.sep), "}\"\n",
+        file = ef, append = TRUE, sep = "")
+    cat("R_LIBS=\"${CALLR_CHILD_R_LIBS:-",
+        paste(libpath, collapse = .Platform$path.sep), "}\"\n",
+        file = ef, append = TRUE, sep = "")
+    cat("R_LIBS_USER=\"${CALLR_CHILD_R_LIBS_USER:-",
+        paste(libpath, collapse = .Platform$path.sep), "}\"\n",
+        file = ef, append = TRUE, sep = "")
   }
 
   c(env_sys, env_user)
@@ -172,5 +195,18 @@ setup_rcmd_binary_and_args <- function(options) {
     options$real_cmdargs <- c("CMD", options$cmd, options$cmdargs)
   }
 
+  options
+}
+
+setup_rscript_binary_and_args <- function(options) {
+
+  if(os_platform() == "windows") {
+    options$bin <- file.path(R.home("bin"), "Rscript.exe")
+
+  } else {
+    options$bin <- file.path(R.home("bin"), "Rscript")
+  }
+
+  options$real_cmdargs <- c(options$script, options$cmdargs)
   options
 }