+++ NEWS.md
@@ -1,4 +1,23 @@
 
+# callr 3.1.0
+
+* New `rscript()` function and `rscript_process` class to execute
+  R scripts via `Rscript` (#40, #81).
+
+* Library paths are now correctly set up for `system()` (and similar)
+  calls from the callr subprocesses (#83, #84).
+
+* Pass `options("repos")` to the child process as is, without checking.
+  Closes #82.
+
+* `r_session$run_with_output()` now returns an S3 object with class
+  `callr_session_result`.
+
+* `r_session$run*()` handle interrupts properly. It tries to interrupt
+  the background process fist, kills it if it is not interruptable,
+  and then re-throws the interrupt condition, going back to the top level
+  prompt if the re-thrown condition is un-caught.
+
 # callr 3.0.0
 
 * New `r_session` class: a background R session you can send commands to